+++ src/module/core/wrap.tpl.html
@@ -1,5 +1,4 @@
 <div class="table-wrap {{ vm.classNames }}">
-    <!--table将会插入table-div内，这个div内不允许有空格-->
     <div class="table-div"></div>
     <span class="text-dreamland"></span>
     {{ vm.configTpl}}
