+++ version/v2.7.x.md
@@ -15,7 +15,8 @@
     - 升级webpack4, 优化构建流程，减少包体积。
     - 升级babel
     - 代码拆分，降低代码偶合度
-
+    - showTh、hideTh传参支持thDOM和thName两种形式
+    - 配置区域增加点击其它区域关闭事件
 
 # v2.7.2
 - 新增
